requirements.txt
Flask
requests
python-dotenv
langchain-openai
langchain-community
streamlit
google-api-python-client
faiss-cpu
google-auth-httplib2
google-auth-oauthlib
APScheduler
celery 
redis
SQLAlchemy